feat(wallet): 3-tab split of WalletSettingsSheet (Design/Details/Connection) (#976)
Restructures the wallet-settings sheet from one long ~85% scroll into a
segmented control with three tabs, completing Option B of the wallet-
settings redesign:

- Design    → the cover-flow WalletCardPicker
- Details   → alias + Lightning address + Save
- Connection→ relay + NWC connection string + QR + CoinOS recovery + migrate
- Remove stays a fixed footer, always visible below the tabs.

Extracts each tab into its own cohesive sub-component (WalletDesignTab /
WalletDetailsTab / WalletConnectionTab) and moves the StyleSheet into
src/styles/WalletSettingsSheet.styles.ts per repo convention. The sheet
drops from 734 to 356 lines and is now pure composition + state.

Segmented-control tabs carry testIDs (wallet-settings-tab-{design,details,
connection}) + accessibilityLabels; Remove button is wallet-settings-remove.
All existing testIDs, handlers, and per-wallet-type logic preserved.
